refactor chakra extension config
