+++ packages/extension-chakra-storefront/src/hooks/use-extension-config.js
@@ -14,7 +14,7 @@ import extensionMeta from '../../extension-meta.json'
 export const useExtensionConfig = () => {
     const extension = useApplicationExtension(extensionMeta.id)
     if (extension === undefined) {
-        throw new Error(`'useExtensionConfig' could find your current application extension instance!`)
+        throw new Error(`'useExtensionConfig' could not find your current application extension instance!`)
     }
     return extension?.getConfig()
 }
